Put checks for access right to own and crud jobs

+++ backend/job-service/src/main/java/com/indireed/jobservice/JobServiceImpl.java
@@ -9,6 +9,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
@@ -22,17 +23,20 @@ public class JobServiceImpl implements JobService {
     private final RestTemplate restTemplate;
 
     @Override
-    public JobDetailDto create(CreateUpdateJobDto request) {
+    public JobDetailDto create(UUID userId, CreateUpdateJobDto request) {
         Job job = new ModelMapper().map(request, Job.class);
+        job.setUserId(userId);
         job = jobRepository.save(job);
         return new ModelMapper().map(job, JobDetailDto.class);
     }
 
     @Override
-    public JobDetailDto update(UUID id, CreateUpdateJobDto request) {
+    public JobDetailDto update(UUID id,UUID userId, CreateUpdateJobDto request) {
         Optional<Job> job = jobRepository.findById(id);
         if (job.isEmpty())
             throw new ResourceNotFoundException("Job not found");
+        if(job.get().getUserId() != userId)
+            throw new AccessDeniedException("You are not allowed to edit this job");
         new ModelMapper().map(request, job.get());
         Job updatedJob = jobRepository.save(job.get());
         return new ModelMapper().map(updatedJob, JobDetailDto.class);
@@ -47,10 +51,12 @@ public class JobServiceImpl implements JobService {
     }
 
     @Override
-    public MessageResponseDto delete(UUID id) {
+    public MessageResponseDto delete(UUID id, UUID userId) {
         Optional<Job> job = jobRepository.findById(id);
         if (job.isEmpty())
             throw new ResourceNotFoundException("Job not found");
+        if(job.get().getUserId() != userId)
+            throw new AccessDeniedException("You are not allowed to delete this job");
         jobRepository.delete(job.get());
         return new MessageResponseDto("Job deleted successfully");
     }
@@ -62,7 +68,7 @@ public class JobServiceImpl implements JobService {
 
         Pageable pageable = PageRequest.of(page, size);
 
-        if (query.isEmpty())
+        if (!query.isEmpty())
             return jobRepository.
                 findAllByPositionIgnoreCaseLikeOrDescriptionIgnoreCaseLike(pageable, query, query)
                     .map(entity -> new ModelMapper().map(entity, JobDetailDto.class));
@@ -72,7 +78,7 @@ public class JobServiceImpl implements JobService {
     }
 
     @Override
-    public MessageResponseDto apply(UUID id) {
+    public MessageResponseDto apply(UUID id, UUID userId) {
         Optional<Job> job = jobRepository.findById(id);
         if (job.isEmpty())
             throw new ResourceNotFoundException("Job not found");
